+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # Revision history for patch
 
-## Unreleased
+## 0.0.3.0
 
 * Create `PatchMapWithPatchingMove` variant which supports moves with a patch.